# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://farm.test:8888/public/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Store new user",
                "description": "Returns user data",
                "operationId": "storeUser",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Authenticate user",
                "description": "Returns user data",
                "operationId": "authenticateUser",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logout user",
                "description": "Returns logged out message",
                "operationId": "logoutUser",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/fields": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Get list of fields",
                "description": "Returns list of fields",
                "operationId": "getFieldList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Fields"
                ],
                "summary": "Store new field",
                "description": "Returns field data",
                "operationId": "storeField",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/fields/{id}": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Get field information",
                "description": "Returns field data",
                "operationId": "getFieldById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Field id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Fields"
                ],
                "summary": "Delete existing field",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteField",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Field id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Fields"
                ],
                "summary": "Update existing field",
                "description": "Returns updated field data",
                "operationId": "updateField",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Field id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/fields/{id}/sensors": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Get field information and its related sensors",
                "description": "Returns field data and its related sensors data",
                "operationId": "getFieldList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/sensors": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Get list of sensors",
                "description": "Returns list of sensors",
                "operationId": "getSensorList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Store new sensor",
                "description": "Returns sensor data",
                "operationId": "storeSensor",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/sensors/{id}": {
            "get": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Get sensor information",
                "description": "Returns sensor data",
                "operationId": "getSensorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sensor id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Delete existing sensor",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteSensor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sensor id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Sensors"
                ],
                "summary": "Update existing sensor",
                "description": "Returns updated sensor data",
                "operationId": "updateSensor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sensor id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get profile",
                "description": "Returns profile data",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update authenticated user",
                "description": "Returns authenticated user data",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}